refactor(stepline): renderized step 1, add inputs and placeholders
