+++ src/app/app.module.ts
@@ -4,21 +4,12 @@ import { AppRoutingModule } from './app-routing.module';
 import { StoreModule } from '@ngrx/store';
 
 import { AppComponent } from './app.component';
-import { modalReducer } from './core/store/modal/modal.reducer';
-
-import { HTTP_INTERCEPTORS } from '@angular/common/http';
-import { AuthInterceptor } from '@services/auth/interceptors';
+import { modalReducer } from '@store/modal/modal.reducer';
 
 @NgModule({
   declarations: [AppComponent],
   imports: [BrowserModule, AppRoutingModule, StoreModule.forRoot({ modal: modalReducer })],
-  providers: [
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: AuthInterceptor,
-      multi: true,
-    },
-  ],
+  providers: [],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
