--- src/app/core/services/auth/guards/index.ts
@@ -1,1 +1,0 @@
-export * from './auth.guard';
