--- src/app/core/services/auth/interceptors/index.ts
@@ -1,1 +1,0 @@
-export * from './auth.interceptor';
