--- src/app/core/services/auth/services/index.ts
@@ -1,5 +1,0 @@
-export * from './auth.service';
-export * from './auth-login';
-export * from './auth-refresh';
-export * from './auth-role';
-export * from './auth-token';
