+++ src/app/pages/home/class/create-task/create-task.component.html
@@ -4,5 +4,8 @@
   <img src="assets/img/empty-task-page.webp" alt="task empty page" />
   <app-button codeIcon="majesticons:plus"></app-button>
   <p>Lista de tarefas vazia, adicione uma nova tarefa</p>
+  <app-step1></app-step1>
+  <app-step2></app-step2>
+  <app-step3></app-step3>
   <app-navbar></app-navbar>
 </app-section>
