--- src/app/pages/home/class/create-task/step1/step1.component.html
@@ -1,1 +1,0 @@
-<app-section> colocar form/input 1 </app-section>
