--- src/app/pages/home/class/create-task/step1/step1.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-step1',
-  templateUrl: './step1.component.html',
-  styleUrl: './step1.component.scss',
-})
-export class Step1Component implements OnInit {
-  ngOnInit(): void {
-    console.log('<app-step1> renderizado.');
-  }
-}
