+++ src/app/shared/components/dynamicView/create-task-views/create-task-views.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { Step1Component } from './step1/step1.component';
+import { Step2Component } from './step2/step2.component';
+import { Step3Component } from './step3/step3.component';
+import { StructuralModule } from '@components/structural/structural.module';
+import { AppFormsModule } from '@components/app-forms/app-forms.module';
+
+@NgModule({
+  declarations: [Step1Component, Step2Component, Step3Component],
+  imports: [CommonModule, StructuralModule, AppFormsModule],
+  exports: [Step1Component, Step2Component, Step3Component],
+})
+export class CreateTaskViewsModule {}
