+++ src/app/shared/components/dynamicView/create-task-views/step1/step1.component.html
@@ -1,0 +1,18 @@
+<app-section>
+  <app-form (onsubmit)="inputData($event)">
+    <app-input-field
+      [required]="true"
+      formControlName="title"
+      label="Título da tarefa"
+      placeholder="Escreva aqui o título da tarefa">
+    </app-input-field>
+    <app-input-field
+      [required]="true"
+      formControlName="description"
+      label="Descrição da tarefa"
+      type="textarea"
+      placeholder="Escreva aqui a desrição da tarefa">
+    </app-input-field>
+    <app-button label="Próximo" (click)="nextStep()"></app-button>
+  </app-form>
+</app-section>
