+++ src/app/shared/components/dynamicView/create-task-views/step1/step1.component.scss
@@ -1,0 +1,3 @@
+button {
+  justify-content: center;
+}
